CHG: Updates for latest ROOT and latest compiler
git-svn-id: https://elves.srl.caltech.edu/svn/nusim/trunk@676 f61b1093-7add-4ab6-ac1b-eb05caf01ebb

+++ src/data/src/NExposureMap.cxx
@@ -168,7 +168,7 @@ bool NExposureMap::WriteExposureMap(int index)
 
 	int bitpix   =  FLOAT_IMG;
 	long naxis    =   2;  /* 2-dimensional image                            */    
-  long naxes[2] = { SkyMapX.size(), SkyMapY.size() };   /* image is 300 pixels wide by 200 rows */
+  long naxes[2] = { long(SkyMapX.size()), long(SkyMapY.size()) };   /* image is 300 pixels wide by 200 rows */
 
   array[0] = (float *)malloc( naxes[0] * naxes[1] * sizeof( float ) );
